+++ src/components/Player/HitterRecent.tsx
@@ -1,5 +1,5 @@
 import { HitterRecentType } from "@customTypes/hitterRecent";
-import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable";
+import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable.style";
 import { createColumnHelper, flexRender, getCoreRowModel, useReactTable } from "@tanstack/react-table";
 import { useState } from "react";
