+++ src/components/Player/HitterTable.tsx
@@ -1,5 +1,5 @@
 import { HitterType } from "@customTypes/playerHitter";
-import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable";
+import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable.style";
 import { createColumnHelper, flexRender, getCoreRowModel, useReactTable } from "@tanstack/react-table";
 import { useState } from "react";
