+++ src/components/Player/PitcherRecent.tsx
@@ -1,5 +1,5 @@
 import { PitcherRecentType } from "@customTypes/pitcherRecent";
-import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable";
+import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable.style";
 import { createColumnHelper, flexRender, getCoreRowModel, useReactTable } from "@tanstack/react-table";
 import { useState } from "react";
