+++ src/components/Player/PitcherTable.tsx
@@ -1,5 +1,5 @@
 import { PitcherType } from "@customTypes/playerPitcher";
-import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable";
+import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable.style";
 import { createColumnHelper, flexRender, getCoreRowModel, useReactTable } from "@tanstack/react-table";
 import { useState } from "react";
