+++ src/components/Player/PitcherTotal.tsx
@@ -1,5 +1,5 @@
 import { PitcherTotalType } from "@customTypes/pitcherTotal";
-import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable";
+import { TableBtd, TableCell, TableHeader } from "@styles/PlayerTable.style";
 import { createColumnHelper, flexRender, getCoreRowModel, useReactTable } from "@tanstack/react-table";
 import { useState } from "react";
