+++ src/types/pitcherTotal.d.ts
